% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootvar.R
\name{bootvar}
\alias{bootvar}
\title{Bootstrap of individuals on a numeric matrix and on a categorical matrix}
\usage{
bootvar(X.quanti = NULL, X.quali = NULL)
}
\arguments{
\item{X.quanti}{a numeric matrix of data, or an object that can be coerced
to such a matrix (such as a numeric vector or a data frame with all numeric columns).}

\item{X.quali}{a categorical matrix of data, or an object that can be
coerced to such a matrix (such as a character vector, a factor or a data
frame with all factor columns).}
}
\description{
Draw a bootstrap sample from X.quanti and a bootstrap sample from X.quali
}
