\name{ClustOfVar-package}
\alias{ClustOfVar-package}
\alias{ClustOfVar}
\docType{package}
\title{
Clustering of variables
}
\description{
Cluster analysis on a set of variables. 
Variables can be quantitative, qualitative or a mixture of both.
Missing values are replaced by means for quantitative variables and by zeros in the indicator matrix for qualitative variables.  
}
\details{
\tabular{ll}{
Package: \tab ClustOfVar\cr
Type: \tab Package\cr
Version: \tab 0.5\cr
Date: \tab 2010-11-06\cr
License: \tab GPL (>=2.0)\cr
LazyLoad: \tab yes\cr
}
}
\author{Marie Chavent <marie.chavent@u-bordeaux2.fr>, Vanessa Kuentz, Benoit Liquet, Jerome Saracco}
\references{
Chavent M., Kuentz V., Liquet B., Saracco J. (2012). ClustOfVar: An R Package for the Clustering Of Variables.\emph{Journal of Statistical Software}, 50(13), 1-16. \cr

Chavent M., Kuentz V., Sarraco J. (2012). Orthogonal rotation in PCAMIX. \emph{Advances in Data Analysis and Classification}, 6(2),131-146.
}
\keyword{ package }
\seealso{
\code{\link{hclustvar}}, \code{\link{kmeansvar}}, \code{\link{stability}}
}

