% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descript.R
\name{descript}
\alias{descript}
\title{Description of a partition of variables}
\usage{
descript(part, rec, matsim = FALSE)
}
\arguments{
\item{part}{a vector with cluster memberships of a partition of variables}

\item{rec}{the element $rec of an object of class clustvar}

\item{matsim}{boolean, if TRUE, the matrices of similarities between variables in same cluster are calculated.}
}
\value{
\item{var}{a list of matrices of squared loadings i.e. for each cluster of variables, the squared loadings on first principal component of PCAmix.
For quantitative variables (resp. qualitative), squared loadings are the squared correlations (resp. the correlation ratios) with the first PC (the cluster center).}
\item{sim}{a list of matrices of similarities i.e. for each cluster, similarities between their variables. 
The similarity between two variables is defined as a square cosine: the square of the Pearson correlation when the two variables are quantitative; 
the correlation ratio when one variable is quantitative and the other one is qualitative; 
the square of the canonical correlation between two sets of dummy variables, when the two variables are qualitative.
\code{sim} is 'NULL' if 'matsim' is 'FALSE'.}
\item{cluster}{a vector of integers indicating the cluster to which each variable is allocated.}
\item{wss}{the within-cluster sum of squares for each cluster: the sum of the correlation ratio (for qualitative variables) and the squared correlation (for quantitative variables) between the variables and the center of the cluster.}
\item{E}{the pourcentage of  homogeneity which is accounted by the partition in k clusters.}
\item{size}{the number of variables in each cluster.}
\item{scores}{a n by k numerical matrix which contains the k cluster centers. The center of a cluster is a synthetic variable: the first principal component calculated by PCAmix. 
The k columns of \code{scores} contain the scores of the n observations units on the first PCs of the k clusters.}
\item{coef}{a list of the coefficients of the linear combinations defining the synthetic variable of each cluster.}
}
\description{
Gives a description of a partition of variables and the scores of the cluster's synthetic variables.
}
\keyword{internal}
