% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mixedVarSim.R
\name{mixedVarSim}
\alias{mixedVarSim}
\title{Similarity between two variables}
\usage{
mixedVarSim(X1, X2)
}
\arguments{
\item{X1}{a vector or a factor}

\item{X2}{a vector or a factor}
}
\description{
Returns the similarity between two quantitative variables, two qualitative
variables or a quantitative variable and a qualitative variable.  The
similarity between two variables is defined as a square cosine: the square
of the Pearson correlation when the two variables are quantitative; the
correlation ratio when one variable is quantitative and the other one is
qualitative; the square of the canonical correlation between two sets of
dummy variables, when the two variables are qualitative.
}
\author{
Marie Chavent \email{Marie.Chavent@u-bordeaux.fr}, Vanessa Kuentz,
Benoit Liquet, Jerome Saracco
}
\keyword{cluster}
\keyword{multivariate}

