\name{print.clustab}
\alias{print.clustab}

\title{
Print a 'clustab' object
}
\description{
This is a method for the function print for objects of the class \code{clustab}. 
}

\usage{
\method{print}{clustab}(x, \dots)
}

\arguments{
  \item{x}{An object of class  \code{clustab} generated by the function \code{\link{stability}}. 
}
  \item{\dots}{
Further arguments to be passed to or from other methods. They are ignored in this function.
}
}

\author{Marie Chavent <marie.chavent@u-bordeaux2.fr>, Vanessa Kuentz, Benoit Liquet, Jerome Saracco}

\seealso{
\code{\link{stability}}
}

\keyword{ print }