\name{print.clustvar}
\alias{print.clustvar}
\title{
Print a 'clustvar' object
}
\description{
This is a method for the function print for objects of the class \code{clustvar}.
}

\usage{
\method{print}{clustvar}(x, \dots)
}

\arguments{
  \item{x}{An object of class  \code{clustvar} generated by the functions \code{\link{cutreevar}} and \code{\link{kmeansvar}}. 
}
  \item{\dots}{
Further arguments to be passed to or from other methods. They are ignored in this function.
}
}

\author{Marie Chavent <marie.chavent@u-bordeaux2.fr>, Vanessa Kuentz, Benoit Liquet, Jerome Saracco}

\seealso{
\code{\link{cutreevar}} , \code{\link{kmeansvar}}
}

\keyword{ print }