\name{print.hclustvar}
\alias{print.hclustvar}
\title{
Print a 'hclustvar' object
}
\description{
This is a method for the function print for objects of the class \code{hclustvar}.
}

\usage{
\method{print}{hclustvar}(x, \dots)
}

\arguments{
  \item{x}{An object of class  \code{hclustvar} generated by the function \code{\link{hclustvar}}. 
}
  \item{\dots}{
Further arguments to be passed to or from other methods. They are ignored in this function.
}
}

\author{Marie Chavent <marie.chavent@u-bordeaux2.fr>, Vanessa Kuentz, Benoit Liquet, Jerome Saracco}

\seealso{
\code{\link{hclustvar}}
}

\keyword{ print }