% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand.R
\name{rand}
\alias{rand}
\title{Rand index between two partitions}
\usage{
rand(P, Q, symmetric = TRUE, adj = TRUE)
}
\arguments{
\item{P}{a factor, e.g., the first partition.}

\item{Q}{a factor, e.g., the second partition.}

\item{symmetric}{a boolean. If FALSE the asymmetrical Rand index is
calculated.}

\item{adj}{a boolean. If TRUE the corrected index is calculated.}
}
\description{
Returns the Rand index, the corrected Rand index or the asymmetrical Rand
index.  The asymmetrical Rand index (corrected or not) measures the
inclusion of a partition P into and partition Q with the number of
clusters in P greater than the number of clusters in Q.
}
\seealso{
\code{\link{stability}}
}
