\name{summary.clustab}
\alias{summary.clustab}
\title{
Summary of a 'clustab' object
}
\description{
This is a method for the function summary for objects of the class \code{clustab}.
}
\usage{
\method{summary}{clustab}(object, \dots)
}

\arguments{
  \item{object}{
An object of class  \code{clustab} generated by the function \code{\link{stability}}. 
}
  \item{\dots}{
Further arguments passed to or from other methods.
}
}

\author{Marie Chavent <marie.chavent@u-bordeaux2.fr>, Vanessa Kuentz, Benoit Liquet, Jerome Saracco}

\seealso{
 \code{\link{stability}}
}

