\name{ClustOfVar-package}
\alias{ClustOfVar-package}
\alias{ClustOfVar}
\docType{package}
\title{
Clustering of variables
}
\description{
Cluster analysis on a set of variables. 
Variables can be quantitative, qualitative or a mixture of both.
Missing values are replaced by means for quantitative variables and by zeros in the indicator matrix for qualitative variables.  
}
\details{
\tabular{ll}{
Package: \tab ClustOfVar\cr
Type: \tab Package\cr
Version: \tab 0.8\cr
Date: \tab 2013-02-06\cr
License: \tab GPL (>=2.0)\cr
LazyLoad: \tab yes\cr
}
}
\author{Marie Chavent <marie.chavent@u-bordeaux2.fr>, Vanessa Kuentz, Benoit Liquet, Jerome Saracco}
\references{
 ClustOfVar: An R Package for the Clustering of Variables. Journal of Statistical Software, Vol. 50, pp. 1-16.
}
\keyword{ package }
\seealso{
\code{\link{hclustvar}}, \code{\link{kmeansvar}}, \code{\link{stability}}
}

