\name{ClustOfVar-package}
\alias{ClustOfVar-package}
\alias{ClustOfVar}
\docType{package}
\title{
Clustering of variables
}
\description{
Cluster analysis on a set of variables. 
Variables can be quantitative, qualitative or a mixture of both.
Missing values are replaced by means for quantitative variables and by zeros in the indicator matrix for qualitative variables.  
}
\details{
\tabular{ll}{
Package: \tab ClustOfVar\cr
Type: \tab Package\cr
Version: \tab 0.5\cr
Date: \tab 2010-11-06\cr
License: \tab GPL (>=2.0)\cr
LazyLoad: \tab yes\cr
}
}
\author{Marie Chavent <marie.chavent@u-bordeaux2.fr>, Vanessa Kuentz, Benoit Liquet, Jerome Saracco}
\references{
Chavent M., Kuentz V., Sarraco J., Clustering of categorical variables around latent variables, Cahiers du GREThA 2010-02, Groupe de Recherche en Economie Theorique et Appliquee, 2010.
Chavent M., Kuentz V., Saracco J., A Partitioning Method for the clustering of Categorical variables, In Classification as a Tool for Research, Hermann Locarek-Junge, Claus Weihs (Eds), Proceedings of the IFCS 2009, Dresden, Springer, 2009. 
}
\keyword{ package }
\seealso{
\code{\link{hclustvar}}, \code{\link{kmeansvar}}, \code{\link{stability}}
}

