% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getnnsvar.R
\name{getnnsvar}
\alias{getnnsvar}
\title{Nearest neighbor of variables based on a score between them}
\usage{
getnnsvar(diss, flag)
}
\arguments{
\item{diss}{a dissimilarity matrix between variables}

\item{flag}{a vector of size \code{p} which indicates if we want to compute nearest
neighbor of variable j (flag[j]=1) or not (flag[j]=0)}
}
\description{
Nearest neighbor of variables based on a score between them
}
\keyword{internal}

