% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.clustvar.R
\name{plot.clustvar}
\alias{plot.clustvar}
\title{Plot scores of variables on synthetic variables.}
\usage{
\method{plot}{clustvar}(x, ...)
}
\arguments{
\item{x}{an object of class \code{clustvar} obtained with \code{cutreevar} or \code{kmeansvar}.}

\item{\dots}{Further arguments to be passed to or from other methods. They
are ignored in this function.}
}
\value{
\item{coord.quanti}{coordinates of quantitative variables belonging to cluster k on the synthetic
 variable associate to the same cluster k}
\item{coord.levels}{coordinates of levels of categorical variables belonging to cluster k on the synthetic
 variable associate to the same cluster k}
}
\description{
Plot scores of variables of cluster k on the synthetic variable associated to the same cluster. For each cluster
two plots are displayed: one for numerical variables (if available in the cluster) and one for levels of categorical
variables (if available in the cluster).
}
\examples{
data(wine)
tree <- hclustvar(data=wine)
tree.cut<-cutree(tree,6)

#plot of scores on synthetic variables
res.plot<-plot(tree.cut)
res.plot$coord.quanti
res.plot$coord.levels
}
\keyword{cluster}
\keyword{multivariate}

