% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.hclustvar.R
\name{plot.hclustvar}
\alias{plot.hclustvar}
\title{Dendrogram of the hierarchy of variables}
\usage{
\method{plot}{hclustvar}(x, type = "tree", which = c(1:2),
  ask = prod(par("mfcol")) < length(which) && dev.interactive(), sub = "",
  ...)
}
\arguments{
\item{x}{an object of class \code{hclustvar}.}

\item{type}{if type="tree" plot of the dendrogram and if type="index"
aggregation levels plot.}

\item{which}{if one of the two plots is required, specify a subset of the
numbers 1:2}

\item{ask}{logical; if TRUE, the user is _ask_ed before each plot.}

\item{sub}{a sub title for the plot.}

\item{\dots}{further arguments passed to or from other methods.}
}
\description{
Dendrogram of the hierarchy of variables resulting from \code{hclustvar}
and aggregation levels plot.
}
\examples{
data(wine)
tree <- hclustvar(data=wine)
plot(tree)

# 2 plots on 1 page
par(mfrow = c(1, 2))
plot(tree)

 # plot just the dendrogram
plot(tree,which=1)
}
\seealso{
\code{\link{hclustvar}}
}
\keyword{cluster}
\keyword{multivariate}

