% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/predict.clustvar.R
\name{predict.clustvar}
\alias{predict.clustvar}
\title{Scores of new objects on the synthetic variables of a given partition}
\usage{
\method{predict}{clustvar}(object, data = NULL, ...)
}
\arguments{
\item{object}{an object of class clustvar}

\item{data}{a dataframe containing data with quantitative and qualitative variables
to be analyzed. Qualitative variable must be coded as factors.}

\item{\dots}{Further arguments to be passed to or from other methods. They
are ignored in this function.}
}
\value{
Returns the matrix of the scores of the new objects on the k syntetic variables of the k-clusters partition given in input.
}
\description{
A partition of variables obtained with kmeansvar or with cutreevar is given in input.
Each cluster of this partition is associated with a synthetic variable which is a linear combination of the variables of the cluster.
The coefficients of these k linear combinations (one for each cluster) are used here to calculate new scores of a objects described in a new dataset (with the same variables).
The output is the matrix of the scores of these new objects on the k synthetic variables.
}
\examples{
data(wine)
n <- nrow(wine)
sub <- 10:20
data.sub <- wine[sub,] #learning sample
part <-kmeansvar(data=data.sub, init=5)
data.t <- wine[-sub, ]
new <- predict(part,data=data.t)
}
\author{
Marie Chavent \email{Marie.Chavent@u-bordeaux.fr}, Vanessa Kuentz, Amaury Labenne, Benoit Liquet, Jerome Saracco
}

