% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.clustab.R
\name{print.clustab}
\alias{print.clustab}
\title{Print a 'clustab' object}
\usage{
\method{print}{clustab}(x, ...)
}
\arguments{
\item{x}{An object of class \code{clustab} generated by the function
\code{\link{stability}}.}

\item{\dots}{Further arguments to be passed to or from other methods. They
are ignored in this function.}
}
\description{
This is a method for the function print for objects of the class
\code{clustab}.
}
\seealso{
\code{\link{stability}}
}
\keyword{print}

