% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/print.clustvar.R
\name{print.clustvar}
\alias{print.clustvar}
\title{Print a 'clustvar' object}
\usage{
\method{print}{clustvar}(x, ...)
}
\arguments{
\item{x}{An object of class \code{clustvar} generated by the functions
\code{\link{cutreevar}} and \code{\link{kmeansvar}}.}

\item{\dots}{Further arguments to be passed to or from other methods. They
are ignored in this function.}
}
\description{
This is a method for the function print for objects of the class
\code{clustvar}.
}
\seealso{
\code{\link{cutreevar}} , \code{\link{kmeansvar}}
}
\keyword{print}

