% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/print.hclustvar.R
\name{print.hclustvar}
\alias{print.hclustvar}
\title{Print a 'hclustvar' object}
\usage{
\method{print}{hclustvar}(x, ...)
}
\arguments{
\item{x}{An object of class \code{hclustvar} generated by the function
\code{\link{hclustvar}}.}

\item{\dots}{Further arguments to be passed to or from other methods. They
are ignored in this function.}
}
\description{
This is a method for the function print for objects of the class
\code{hclustvar}.
}
\seealso{
\code{\link{hclustvar}}
}
\keyword{print}

