% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.clustab.R
\name{summary.clustab}
\alias{summary.clustab}
\title{Summary of a 'clustab' object}
\usage{
\method{summary}{clustab}(object, ...)
}
\arguments{
\item{object}{An object of class \code{clustab} generated by the function
\code{\link{stability}}.}

\item{\dots}{Further arguments passed to or from other methods.}
}
\description{
This is a method for the function summary for objects of the class
\code{clustab}.
}
\seealso{
\code{\link{stability}}
}

